--- a/src/algorithms.jl
+++ b/src/algorithms.jl
@@ -56,6 +56,7 @@ struct DGLDDRK73_C <: OrdinaryDiffEqAlgorithm end
 struct DGLDDRK84_C <: OrdinaryDiffEqAlgorithm end
 struct CarpenterKennedy2N54 <: OrdinaryDiffEqAlgorithm end
 struct ParsaniKetchesonDeconinck3S184 <: OrdinaryDiffEqAlgorithm end
+struct ParsaniKetchesonDeconinck3S205 <: OrdinaryDiffEqAlgorithm end
 struct ORK256 <: OrdinaryDiffEqAlgorithm end
 struct SSPRK22{StageLimiter,StepLimiter} <: OrdinaryDiffEqAlgorithm
   stage_limiter!::StageLimiter