--- a/test/ode/ode_cache_tests.jl
+++ b/test/ode/ode_cache_tests.jl
@@ -2,7 +2,7 @@ using OrdinaryDiffEq, DiffEqBase, DiffEqCallbacks, Test
 using Random
 CACHE_TEST_ALGS = [Euler(),Midpoint(),RK4(),SSPRK22(),SSPRK33(),SSPRK53(), SSPRK53_2N1(), SSPRK53_2N2(),
   SSPRK63(),SSPRK73(),SSPRK83(),SSPRK432(),SSPRK932(),SSPRK54(),SSPRK104(),
-  CarpenterKennedy2N54(),ParsaniKetchesonDeconinck3S184(),
+  CarpenterKennedy2N54(),ParsaniKetchesonDeconinck3S184(),ParsaniKetchesonDeconinck3S205(),
   BS3(),BS5(),DP5(),DP5Threaded(),DP8(),Feagin10(),Feagin12(),Feagin14(),TanYam7(),
   Tsit5(),TsitPap8(),Vern6(),Vern7(),Vern8(),Vern9(),OwrenZen3(),OwrenZen4(),OwrenZen5()]