--- a/test/ode/ode_dense_tests.jl
+++ b/test/ode/ode_dense_tests.jl
@@ -176,13 +176,15 @@ regression_test(SSPRK54(), 3.5e-5, 5.5e-5)
 # SSPRK104
 regression_test(SSPRK104(), 1.5e-5, 3e-5)
 
+println("Low Storage RKs")
+
+# CarpenterKennedy2N54
+regression_test(CarpenterKennedy2N54(), 3.0e-5, 5.0e-5)
+
 println("RKs")
 
 # RK4
 regression_test(RK4(), 4.5e-5, 1e-4)
-
-# CarpenterKennedy2N54
-regression_test(CarpenterKennedy2N54(), 3.0e-5, 5.0e-5)
 
 # DP5
 regression_test(DP5(), 5e-6, 1e-5; test_diff1 = true, nth_der = 4, dertol = 1e-14)