At first, I have created some new files for low storage methods as suggested in https://github.com/JuliaDiffEq/OrdinaryDiffEq.jl/pull/599#issuecomment-460178896. The cahces for `CarpenterKennedy2N54` have been changed such that a general low-storage implementation of 2N schemes is used. This does not seem to impact the performance negatively:

- This PR
  ```julia
  julia> using OrdinaryDiffEq, BenchmarkTools

  julia> using DiffEqProblemLibrary.ODEProblemLibrary: importodeproblems; importodeproblems()

  julia> import DiffEqProblemLibrary.ODEProblemLibrary: prob_ode_linear, prob_ode_2Dlinear

  julia> @benchmark solve(prob_ode_linear, CarpenterKennedy2N54(), dt=1/2^7, save_everystep=false)
  BenchmarkTools.Trial:
    memory estimate:  4.92 KiB
    allocs estimate:  79
    --------------
    minimum time:     6.273 μs (0.00% GC)
    median time:      6.492 μs (0.00% GC)
    mean time:        6.972 μs (5.61% GC)
    maximum time:     394.342 μs (97.32% GC)
    --------------
    samples:          10000
    evals/sample:     5

  julia> @benchmark solve(prob_ode_2Dlinear, CarpenterKennedy2N54(), dt=1/2^7, save_everystep=false)
  BenchmarkTools.Trial:
    memory estimate:  6.34 KiB
    allocs estimate:  92
    --------------
    minimum time:     28.628 μs (0.00% GC)
    median time:      29.820 μs (0.00% GC)
    mean time:        32.162 μs (1.94% GC)
    maximum time:     2.361 ms (96.59% GC)
    --------------
    samples:          10000
    evals/sample:     1
  ```

- Current master
  ```julia
  julia> using OrdinaryDiffEq, BenchmarkTools

  julia> using DiffEqProblemLibrary.ODEProblemLibrary: importodeproblems; importodeproblems()

  julia> import DiffEqProblemLibrary.ODEProblemLibrary: prob_ode_linear, prob_ode_2Dlinear

  julia> @benchmark solve(prob_ode_linear, CarpenterKennedy2N54(), dt=1/2^7, save_everystep=false)
  BenchmarkTools.Trial:
    memory estimate:  4.92 KiB
    allocs estimate:  79
    --------------
    minimum time:     6.860 μs (0.00% GC)
    median time:      7.057 μs (0.00% GC)
    mean time:        7.557 μs (5.19% GC)
    maximum time:     402.007 μs (97.16% GC)
    --------------
    samples:          10000
    evals/sample:     5

  julia> @benchmark solve(prob_ode_2Dlinear, CarpenterKennedy2N54(), dt=1/2^7, save_everystep=false)
  BenchmarkTools.Trial:
    memory estimate:  6.34 KiB
    allocs estimate:  92
    --------------
    minimum time:     29.327 μs (0.00% GC)
    median time:      30.412 μs (0.00% GC)
    mean time:        31.475 μs (1.94% GC)
    maximum time:     2.606 ms (97.23% GC)
    --------------
    samples:          10000
    evals/sample:     1
  ```
Then, I've added the "big monsters" of [Parsani, Ketcheson, Deconinck (2013)](https://epubs.siam.org/doi/10.1137/120885899), cf. #493. The schemes with less stages are still missing but can be implemented similarly (e.g. in this PR if the general setup is acceptable).