@ranocha 2019-02-06
It's a bit strange that the convergence tests for `ParsaniKetchesonDeconinck3S205()` fail on Travis since they pass locally. However, they are very fragile and even some minor changes to the parameters (`dts`) changed the estimated order of convergence significantly for me. I will experiment a bit more.