@ChrisRackauckas 2019-02-06
Show a plot. You might need to change dts since at 5th order you are hitting floating point truncation error if it's the same as the 3rd order dts.