@ranocha 2019-02-06
Changing `dts` does either result in other estimates for the order of convergence or in errors:
```julia
julia> dts = 1 ./ 2 .^(4:-1:1); sim = test_convergence(dts, prob_ode_nonlinear, alg); display(sim.𝒪est); plot(sim)
ERROR: KeyError: key :l∞ not found
Stacktrace:
 [1] getindex at ./dict.jl:478 [inlined]
 [2] #52 at ./none:0 [inlined]
 [3] iterate at ./generator.jl:47 [inlined]
 [4] collect_to!(::Array{Float64,1}, ::Base.Generator{Array{ODESolution{Float64,1,Array{Float64,1},Array{Float64,1},Dict{Symbol,Float64},Array{Float64,1},Array{Array{Float64,1},1},ODEProblem{Float64,Tuple{Float64,Float64},false,Nothing,ODEFunction{false,getfield(Main, Symbol("##7#8")),LinearAlgebra.UniformScaling{Bool},getfield(DiffEqBase, Symbol("##166#172")){getfield(Main, Symbol("##7#8"))},Nothing,Nothing,Nothing,Nothing,Nothing,Nothing,Nothing},Nothing,DiffEqBase.StandardODEProblem},ParsaniKetchesonDeconinck3S205,OrdinaryDiffEq.InterpolationData{ODEFunction{false,getfield(Main, Symbol("##7#8")),LinearAlgebra.UniformScaling{Bool},getfield(DiffEqBase, Symbol("##166#172")){getfield(Main, Symbol("##7#8"))},Nothing,Nothing,Nothing,Nothing,Nothing,Nothing,Nothing},Array{Float64,1},Array{Float64,1},Array{Array{Float64,1},1},OrdinaryDiffEq.ParsaniKetchesonDeconinck3S205ConstantCache{Float64,Float64}}},1},getfield(DiffEqDevTools, Symbol("##52#53")){Symbol}}, ::Int64, ::Int64) at ./array.jl:651
 [5] collect_to_with_first! at ./array.jl:630 [inlined]
 [6] collect(::Base.Generator{Array{ODESolution{Float64,1,Array{Float64,1},Array{Float64,1},Dict{Symbol,Float64},Array{Float64,1},Array{Array{Float64,1},1},ODEProblem{Float64,Tuple{Float64,Float64},false,Nothing,ODEFunction{false,getfield(Main, Symbol("##7#8")),LinearAlgebra.UniformScaling{Bool},getfield(DiffEqBase, Symbol("##166#172")){getfield(Main, Symbol("##7#8"))},Nothing,Nothing,Nothing,Nothing,Nothing,Nothing,Nothing},Nothing,DiffEqBase.StandardODEProblem},ParsaniKetchesonDeconinck3S205,OrdinaryDiffEq.InterpolationData{ODEFunction{false,getfield(Main, Symbol("##7#8")),LinearAlgebra.UniformScaling{Bool},getfield(DiffEqBase, Symbol("##166#172")){getfield(Main, Symbol("##7#8"))},Nothing,Nothing,Nothing,Nothing,Nothing,Nothing,Nothing},Array{Float64,1},Array{Float64,1},Array{Array{Float64,1},1},OrdinaryDiffEq.ParsaniKetchesonDeconinck3S205ConstantCache{Float64,Float64}}},1},getfield(DiffEqDevTools, Symbol("##52#53")){Symbol}}) at ./array.jl:611
 [7] #ConvergenceSimulation#51(::Dict{Symbol,Array{Float64,1}}, ::Nothing, ::Type, ::Array{ODESolution{Float64,1,Array{Float64,1},Array{Float64,1},Dict{Symbol,Float64},Array{Float64,1},Array{Array{Float64,1},1},ODEProblem{Float64,Tuple{Float64,Float64},false,Nothing,ODEFunction{false,getfield(Main, Symbol("##7#8")),LinearAlgebra.UniformScaling{Bool},getfield(DiffEqBase, Symbol("##166#172")){getfield(Main, Symbol("##7#8"))},Nothing,Nothing,Nothing,Nothing,Nothing,Nothing,Nothing},Nothing,DiffEqBase.StandardODEProblem},ParsaniKetchesonDeconinck3S205,OrdinaryDiffEq.InterpolationData{ODEFunction{false,getfield(Main, Symbol("##7#8")),LinearAlgebra.UniformScaling{Bool},getfield(DiffEqBase, Symbol("##166#172")){getfield(Main, Symbol("##7#8"))},Nothing,Nothing,Nothing,Nothing,Nothing,Nothing,Nothing},Array{Float64,1},Array{Float64,1},Array{Array{Float64,1},1},OrdinaryDiffEq.ParsaniKetchesonDeconinck3S205ConstantCache{Float64,Float64}}},1}, ::Array{Float64,1}) at /.../.julia/packages/DiffEqDevTools/DnQ2e/src/convergence.jl:19
 [8] Type at ./none:0 [inlined]
 [9] #test_convergence#72(::Bool, ::Bool, ::Base.Iterators.Pairs{Union{},Union{},Tuple{},NamedTuple{(),Tuple{}}}, ::Function, ::Array{Float64,1}, ::ODEProblem{Float64,Tuple{Float64,Float64},false,Nothing,ODEFunction{false,getfield(Main, Symbol("##7#8")),LinearAlgebra.UniformScaling{Bool},getfield(DiffEqBase, Symbol("##166#172")){getfield(Main, Symbol("##7#8"))},Nothing,Nothing,Nothing,Nothing,Nothing,Nothing,Nothing},Nothing,DiffEqBase.StandardODEProblem}, ::ParsaniKetchesonDeconinck3S205) at /.../.julia/packages/DiffEqDevTools/DnQ2e/src/convergence.jl:98
 [10] test_convergence(::Array{Float64,1}, ::ODEProblem{Float64,Tuple{Float64,Float64},false,Nothing,ODEFunction{false,getfield(Main, Symbol("##7#8")),LinearAlgebra.UniformScaling{Bool},getfield(DiffEqBase, Symbol("##166#172")){getfield(Main, Symbol("##7#8"))},Nothing,Nothing,Nothing,Nothing,Nothing,Nothing,Nothing},Nothing,DiffEqBase.StandardODEProblem}, ::ParsaniKetchesonDeconinck3S205) at /.../.julia/packages/DiffEqDevTools/DnQ2e/src/convergence.jl:95
 [11] top-level scope at none:0
```