@ChrisRackauckas 2019-02-06
The first one is in the right range. I wonder why it wobbles. When it fails, how close is it?