@ranocha 2019-02-06
On Travis, the estimated order of convergence if `4.46` (instead of `4.86` for me) while we want `5`.