@ChrisRackauckas 2019-02-06
Travis might not have hardware FMA so `@muladd` is making your local computer's computations more numerically stable at low `dt`. We've seen this kind of issue in DelayDiffEq.