@ranocha 2019-02-06
The tests on Travis passed using the new tolerances. Now, I've added some of the other schemes.