@ChrisRackauckas 2019-02-06
>By the way, it might be nice to create a`GROUP=AlgConvergence_III` for Travis CI in order to reduce the build/test time.

Yes, we've probably reached this point.